fix(web): unblock Vercel build and ship dashboard E2E loop
- Remove styled-jsx from ContactForm and HeroSection; move styles to globals.css
- Extract contact-form validation and add landing style safety tests
- Refactor dashboard into DashboardSplitView with persisted store and 4-step workflow
- Fix backend dispatch to use auto-extracted events instead of empty request.events
- Stub server-only in vitest; extend agent pipeline timeout to 210s

Verified: apps/web 112/112 tests pass, npm run build succeeds.
