+++ apps/web/package.json
@@ -7,6 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "eslint src",
+    "test": "vitest run",
     "analyze": "next experimental-analyze --output"
   },
   "dependencies": {
@@ -25,7 +26,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "lucide-react": "^1.21.0",
-    "next": "^16.2.7",
+    "next": "^16.2.9",
     "next-auth": "^4.24.14",
     "openai": "^6.43.0",
     "react": "^18",
@@ -44,7 +45,7 @@
     "@tailwindcss/postcss": "^4.3.1",
     "autoprefixer": "^10.0.1",
     "eslint": "^9",
-    "eslint-config-next": "^16.2.7",
+    "eslint-config-next": "^16.2.9",
     "postcss": "^8.5.10",
     "tailwindcss": "^4.3.1",
     "typescript": "^5",
