+++ apps/web/src/app/ContactForm.tsx
@@ -3,6 +3,11 @@
 import { useCallback, useState } from 'react';
 
 import { CONTACT_EMAIL } from '@/lib/constants';
+import {
+  CONTACT_MESSAGE_MAX,
+  CONTACT_NAME_MAX,
+  validateContactForm,
+} from '@/lib/contact-form-validation';
 
 const TEAL = '#6af2de';
 const TEAL_DEEP = '#10b7a5';
@@ -19,28 +24,6 @@ const USE_CASES = [
   'Other video-to-action workflow',
 ];
 
-const MESSAGE_MAX = 1000;
-const NAME_MAX = 100;
-
-function isEmail(value: string) {
-  return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(value);
-}
-
-function normalizeVideoUrl(value: string) {
-  if (!value) return '';
-  return /^https?:\/\//i.test(value) ? value : `https://${value}`;
-}
-
-function isYouTube(value: string) {
-  if (!value) return true;
-  try {
-    const u = new URL(value);
-    return ['youtube.com', 'www.youtube.com', 'youtu.be', 'm.youtube.com'].includes(u.hostname);
-  } catch {
-    return false;
-  }
-}
-
 type FormStatus = { kind: 'idle' | 'pending' | 'error'; text: string };
 
 export default function ContactForm() {
@@ -54,44 +37,14 @@ export default function ContactForm() {
   const handleSubmit = useCallback(
     (e: React.FormEvent) => {
       e.preventDefault();
-      const n = name.trim();
-      const em = email.trim();
-      const uc = useCase.trim();
-      const v = normalizeVideoUrl(videoUrl.trim());
-      const msg = message.trim();
-
-      if (!n || !em || !uc || !msg) {
-        setStatus({ kind: 'error', text: 'Please fill out name, email, use case, and the short note.' });
-        return;
-      }
-      if (n.length > NAME_MAX || msg.length > MESSAGE_MAX) {
-        setStatus({
-          kind: 'error',
-          text: `Keep the name under ${NAME_MAX} characters and the note under ${MESSAGE_MAX.toLocaleString()} characters.`,
-        });
-        return;
-      }
-      if (!isEmail(em)) {
-        setStatus({ kind: 'error', text: 'Please enter a valid email address.' });
-        return;
-      }
-      if (!isYouTube(v)) {
-        setStatus({ kind: 'error', text: 'If you include a sample video, use a YouTube URL.' });
+      const result = validateContactForm({ name, email, useCase, videoUrl, message });
+      if (!result.ok) {
+        setStatus({ kind: 'error', text: result.error });
         return;
       }
 
-      const subject = encodeURIComponent('UVAI inbound: ' + uc);
-      const body = encodeURIComponent(
-        [
-          'Name: ' + n,
-          'Email: ' + em,
-          'Use case: ' + uc,
-          v ? 'Sample video: ' + v : 'Sample video: not provided',
-          '',
-          'Note:',
-          msg,
-        ].join('\n'),
-      );
+      const subject = encodeURIComponent(result.mailto.subject);
+      const body = encodeURIComponent(result.mailto.body);
 
       setStatus({ kind: 'pending', text: 'Trying to open your email app...' });
       window.location.href = `mailto:${CONTACT_EMAIL}?subject=${subject}&body=${body}`;
@@ -117,11 +70,11 @@ export default function ContactForm() {
             name="name"
             autoComplete="name"
             placeholder="Your name"
-            maxLength={NAME_MAX}
+            maxLength={CONTACT_NAME_MAX}
             required
             value={name}
             onChange={(e) => setName(e.target.value)}
-            className="form-input"
+            className="contact-form-input"
           />
         </Field>
         <Field label="Email" htmlFor="email">
@@ -134,7 +87,7 @@ export default function ContactForm() {
             required
             value={email}
             onChange={(e) => setEmail(e.target.value)}
-            className="form-input"
+            className="contact-form-input"
           />
         </Field>
         <Field label="Use case" htmlFor="use_case">
@@ -144,7 +97,7 @@ export default function ContactForm() {
             required
             value={useCase}
             onChange={(e) => setUseCase(e.target.value)}
-            className="form-input"
+            className="contact-form-input"
           >
             <option value="">Choose one</option>
             {USE_CASES.map((opt) => (
@@ -164,7 +117,7 @@ export default function ContactForm() {
             aria-describedby="video_url_hint"
             value={videoUrl}
             onChange={(e) => setVideoUrl(e.target.value)}
-            className="form-input"
+            className="contact-form-input"
           />
           <span id="video_url_hint" className="text-xs" style={{ color: FAINT }}>
             Optional. YouTube link, with or without https://.
@@ -175,12 +128,12 @@ export default function ContactForm() {
             id="message"
             name="message"
             placeholder="Example: turn product demo videos into API docs and tickets."
-            maxLength={MESSAGE_MAX}
+            maxLength={CONTACT_MESSAGE_MAX}
             required
             rows={4}
             value={message}
             onChange={(e) => setMessage(e.target.value)}
-            className="form-input"
+            className="contact-form-input"
           />
         </Field>
         <p
@@ -206,40 +159,6 @@ export default function ContactForm() {
           Send request
         </button>
       </form>
-
-      <style jsx>{`
-        .form-input {
-          width: 100%;
-          padding: 0.7rem 0.9rem;
-          border-radius: 0.55rem;
-          background: rgba(14, 14, 19, 0.7);
-          border: 1px solid ${BORDER};
-          color: ${INK};
-          font-size: 0.9rem;
-          transition: border-color 0.2s, box-shadow 0.2s;
-        }
-        .form-input::placeholder {
-          color: rgba(248, 245, 253, 0.3);
-        }
-        .form-input:focus {
-          outline: none;
-          border-color: ${TEAL};
-          box-shadow: 0 0 0 3px rgba(106, 242, 222, 0.15);
-        }
-        select.form-input {
-          appearance: none;
-          background-image: linear-gradient(45deg, transparent 50%, ${TEAL} 50%),
-            linear-gradient(135deg, ${TEAL} 50%, transparent 50%);
-          background-position: calc(100% - 18px) 50%, calc(100% - 13px) 50%;
-          background-size: 5px 5px, 5px 5px;
-          background-repeat: no-repeat;
-          padding-right: 2.2rem;
-        }
-        textarea.form-input {
-          resize: vertical;
-          min-height: 100px;
-        }
-      `}</style>
     </article>
   );
 }
