+++ apps/web/src/app/globals.css
@@ -574,6 +574,66 @@ pre {
   }
 }
 
+@keyframes landing-marquee {
+  from {
+    transform: translateX(0);
+  }
+  to {
+    transform: translateX(-50%);
+  }
+}
+
+.animate-marquee {
+  animation: landing-marquee 22s linear infinite;
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .animate-marquee {
+    animation: none;
+  }
+}
+
+.contact-form-input {
+  width: 100%;
+  padding: 0.7rem 0.9rem;
+  border-radius: 0.55rem;
+  background: rgba(14, 14, 19, 0.7);
+  border: 1px solid rgba(72, 71, 77, 0.18);
+  color: #f8f5fd;
+  font-size: 0.9rem;
+  transition:
+    border-color 0.2s,
+    box-shadow 0.2s;
+}
+
+.contact-form-input::placeholder {
+  color: rgba(248, 245, 253, 0.3);
+}
+
+.contact-form-input:focus {
+  outline: none;
+  border-color: #6af2de;
+  box-shadow: 0 0 0 3px rgba(106, 242, 222, 0.15);
+}
+
+select.contact-form-input {
+  appearance: none;
+  background-image:
+    linear-gradient(45deg, transparent 50%, #6af2de 50%),
+    linear-gradient(135deg, #6af2de 50%, transparent 50%);
+  background-position:
+    calc(100% - 18px) 50%,
+    calc(100% - 13px) 50%;
+  background-size: 5px 5px, 5px 5px;
+  background-repeat: no-repeat;
+  padding-right: 2.2rem;
+}
+
+textarea.contact-form-input {
+  resize: vertical;
+  min-height: 100px;
+}
+
 /* ============================================
    Stitch Dashboard Animations
    ============================================ */
