+++ apps/web/src/store/dashboard-types.ts
@@ -1,0 +1,62 @@
+import type { ExtractedEvent, AgentExecution } from '@/lib/types';
+
+export interface PipelineResult {
+  live_url: string | null;
+  github_repo: string | null;
+  build_status: string;
+  code_generation: {
+    framework: string;
+    files_created: string[];
+    entry_point: string;
+  } | null;
+  deployment: {
+    status: string;
+    platforms: string[];
+    urls: Record<string, string>;
+  } | null;
+}
+
+export interface Action {
+  title: string;
+  description: string;
+  category: string;
+  estimatedMinutes?: number | null;
+}
+
+export type PipelineMode = 'live' | 'serverless' | 'fallback' | 'handoff';
+
+export interface Video {
+  id: string;
+  title: string;
+  url: string;
+  status: 'processing' | 'complete' | 'failed';
+  progress: number;
+  /** How analysis ran: backend SSE, Gemini, legacy /api/video, or offline handoff. */
+  pipelineMode?: PipelineMode;
+  thumbnail?: string;
+  duration?: string;
+  processedAt?: string;
+  transcript?: string;
+  events?: ExtractedEvent[];
+  agents?: AgentExecution[];
+  pipelineResult?: PipelineResult;
+  insights?: {
+    summary: string;
+    actions: Action[];
+    sentiment: string;
+    topics: string[];
+  };
+}
+
+export interface Activity {
+  time: string;
+  event: string;
+  type: 'success' | 'info' | 'error';
+}
+
+export interface SearchResult {
+  start: number;
+  duration: number;
+  text: string;
+  score: number;
+}
