+++ apps/web/src/test/server-only.ts
@@ -1,0 +1,2 @@
+// Vitest stub — `server-only` throws when imported outside Server Components.
+export {};
