+++ package.json
@@ -28,7 +28,7 @@
     "vitest": "^4.1.9"
   },
   "overrides": {
-    "next": ">=16.2.6",
+    "next": ">=16.2.9",
     "nth-check": "^2.0.1",
     "postcss": "^8.5.15",
     "undici": "^6.22.0",
